Add support for private cluster with public endpoint disabled
Given a private cluster with public endpoint disabled, the public
endpoint still exists but protected by a firewall rule.
Similar to MAN, CloudRun is able to bypass that rule. Since GKE
people decide to keep this bug, we'll utilize this as a feature.

Tested with manual curl command to call Thetis runIstiod API.

BUG: b/193902968
Change-Id: I3fd3338d82992ec0c51028ab952778ad21343b60