Fixes to TaaA Image structures.
There were issues with how the files were copied into the TaaA
Docker images. This causes file paths to be wrong.

Fixed bugs within the code running the intergration tests.
The integration-tests image is still failing because there are still
issues with their execution, but the test-artifact code is at least now correct.

Added a program to locally run the TaaA docker image for testing.

Change-Id: I2b3a42adb240ebd5eb5e6079a570656e4320a202