Enable stackdriver monitoring in mcp, include sd view map
Previously, stackdriver monitoring wasn't enabled in the new
mcp-specific init. Enable it.

Also include the sd view map so that all the stackdriver views are
reported.

Unrelated: Also require authorization for debug endpoints, to match
the behavior of the script (including for convenience, but it's a big
security hole if we don't include this change).

Change-Id: I8533c202b3e6e286fbea3d757974e937862edd5c