Fixes to get integration test binaries to run.
The Pilot test now runs without crashing issues!
Pilot test is failing on execution, so that still needs fixing.
Junit generated by TaaA lib also is wrong and needs fixing.

Was necessary to create a second images with the same base OS as the
TaaA image. The binaries need to be consistent or runtime errors with
librarires would occur during initialization due to differing versions.

Will later move the command to build using an image to TaaA lib.

BUG=195026676

Change-Id: Ibd1fe59fc47817d1b2ea9cf6239b8a82722c33f2