Support using the rookery config file from TRAC config
This change adds a new --trac-cluster-version flag. If it's specified as
a value >=0, the config file generated from TRAC will be used. For
example, if it's 0, version n will be used, if it's 1, version n-1 will
be used. The default value is set as -1 which means the current custom
rookery config in the ASM repo will be used.

Change-Id: I13fa7099e5658b85473ff6897de6ac5d8d2db5af