fix race causing ingress sa overrwrite
In some of our flakes, especially TestAuthzCheck, I notice we have 2 ingress instances that are nearly identical,
but one has ImagePullBackoff. This happens because it didn't come up before we hit the line of code changed in this PR.
We manually create a valid service account which later gets overwritten because the check wasn't looking for the right name.

Change-Id: Iff04aff1de75394a9e5d1dd4a69d6eb3ef75da80