mcp: use native Telemetry, dropping EnvoyFilter
This changes MCP's telemetry implementation to use native code rather
than EnvoyFilters. This should be invisible to users - they could not
see the EnvoyFilters before, and the new native could should give
identical behavior (SD access logs enable, prometheus+sd metrics
enabled).

If users wish to 'enable optional features', the mechanisms currently
documented will continue to work.

The one area this may impact users is if they are using unsupported
EnvoyFilters that expect to very exactly interact with the builtin ones.
This is not supported, and I hope anyone doing this should be explicitly
REMOVEing the filter, not trying to do something like occluding the
EnvoyFilter name itself to remove it.

Change-Id: Id104c75a01c426e13c007a82d167b823427c8966