Convert the GKE upgrade jobs to be n-2->n-1->n.
Context: This CL will make the upgradeClusterVersion flag a comma seperated list of versions that the cluster should be upgraded too. In between GKE Upgrades we will also sleep for a short amount of time to ensure that the upgrade completes (for GKE ON GCP this time is set to 10 minutes).

Change-Id: I3e2f77c09965a3027b294cc6d8a2a953d1a05eff