Avoid calling `json.Unmarshall` on pointer to pointer
When we do this:
```
x := &foo{}
json.Unmarshal("null", &x)
```

`x` may now be `nil`. This often is unexpected and will lead to panics
later on that expect `x` to not be nil.

This was found by fuzzing. I found other usages by hacking json.Unmarshal:

```
	s := rv.Type().String()
	if len(s) > 2 && s[0] == '*' && s[1] == '*' {
		if !(s == "**types.NetConf" || s=="**jsonpatch.partialDoc" || s == "**configfile.ConfigFile" ) {
			panic(s)
		}
	}
```

With a few types from external libraries with this issue ignored

Change-Id: I1927ba560aea72d1b5ba05541e150550436f835c