1.12 cherrypick: drop XDS concurrency
This was originally merged into 1.10 and 1.11, causing a (unreported) regression for 1.12.
Adding this to 1.12 as well. Original PR was https://gke-internal-review.googlesource.com/c/istio/istio/+/330133.

For 1.13/master we have https://github.com/istio/istio/pull/37117#issuecomment-1027613736, but it only changes on var. Will send a PR for the other.

Change-Id: I38b8ff9eed45046fe56ac5d333ab5f9820eb6e81