use offline mode to install MCP with AFC
This is because we need to test the newly built images in CI. This is
done by using an annotation in the CPRs. However, that annotation should
not be user-facing because it could create confusion both for the
customers and us. This changes the behavior to patching that annotation
in test infra instead of relying on asmcli

Change-Id: I0e41cbdcf7031ad96ac3a4303da8a969eec3843c